Migrated testcases of template module to py.test
